+++ src/Oro/Bundle/ApiBundle/Tests/Functional/RestPlain/InvalidRequestDataTest.php
@@ -8,6 +8,31 @@ use Symfony\Component\HttpFoundation\Response;
 
 class InvalidRequestDataTest extends RestPlainApiTestCase
 {
+    public function testEmptyJsonInRequestData()
+    {
+        $response = $this->request(
+            'POST',
+            $this->getUrl(
+                $this->getListRouteName(),
+                ['entity' => $this->getEntityType(TestProduct::class)]
+            ),
+            [],
+            [],
+            ''
+        );
+        self::assertResponseStatusCodeEquals($response, Response::HTTP_BAD_REQUEST);
+        self::assertResponseContentTypeEquals($response, self::JSON_CONTENT_TYPE);
+        self::assertEquals(
+            [
+                [
+                    'title'  => 'request data constraint',
+                    'detail' => 'The request data should not be empty'
+                ]
+            ],
+            self::jsonToArray($response->getContent())
+        );
+    }
+
     public function testInvalidJsonInRequestData()
     {
         $response = $this->request(
@@ -23,7 +48,11 @@ class InvalidRequestDataTest extends RestPlainApiTestCase
         self::assertResponseStatusCodeEquals($response, Response::HTTP_BAD_REQUEST);
         self::assertResponseContentTypeEquals($response, 'application/json');
         self::assertEquals(
-            ['code' => 400, 'message' => 'Bad Request'],
+            [
+                'code'    => 400,
+                'message' => 'Invalid json message received.'
+                    . ' Parsing error in [1:22]. Expected \'null\'. Got: test'
+            ],
             self::jsonToArray($response->getContent())
         );
     }