+++ CHANGELOG.md
@@ -4,12 +4,20 @@ The current file describes significant changes in the code that may affect the u
 
 ## 4.1.0-beta
 
+#### ActivityBundle
+* The DIC tag `oro_activity.activity_entity_delete_handler` was removed.
+  Use decoration of `oro_activity.activity_entity_delete_handler_extension` service to instead.
+* The interface `Oro\Bundle\ActivityBundle\Entity\Manager\ActivityEntityDeleteHandlerInterface` was removed.
+  Use `Oro\Bundle\ActivityBundle\Handler\ActivityEntityDeleteHandlerExtensionInterface` instead.
+
 #### ApiBundle
 * The section `relations` was removed from `Resources/config/oro/api.yml`. The action `get_relation_config` that
   was responsible to process this section was removed as well.
   This section was not used to build API that conforms JSON:API specification that is the main API type.
   In case if you need a special configuration for "plain" REST API, you can define it in
   `Resources/config/oro/api_plain.yml` configuration files or create a processor for the `get_config` action.
+* The `delete_handler` configuration option was removed.
+  The `Oro\Bundle\EntityBundle\Handler\EntityDeleteHandlerRegistry` class is used to get the deletion handler instead.
 
 #### AssetBundle
 * The new feature, [Hot Module Replacement (HMR or Hot Reload) enabled for SCSS](./src/Oro/Bundle/AssetBundle/Resources/doc/index.md#hot-module-replacement-hmr-or-hot-reload-for-scss). To enable HMR for custom CSS links, please [follow the documentation](./src/Oro/Bundle/AssetBundle/Resources/doc/index.md#enable-for-css-links).
@@ -23,6 +31,13 @@ The current file describes significant changes in the code that may affect the u
   was renamed to `Oro\Bundle\SecurityBundle\Authentication\Token\OrganizationAwareTokenInterface`.
   Also methods `getOrganizationContext` and `setOrganizationContext` were renamed to
   `getOrganization` and `setOrganization`.
+* The class `Oro\Bundle\SecurityBundle\Exception\ForbiddenException` was removed.
+  Use `Symfony\Component\Security\Core\Exception\AccessDeniedException` instead.
+
+#### SoapBundle
+* The interface `Oro\Bundle\SoapBundle\Handler\DeleteHandlerInterface` was replaced with
+  `Oro\Bundle\EntityBundle\Handler\EntityDeleteHandlerInterface`
+  and `Oro\Bundle\EntityBundle\Handler\EntityDeleteHandlerExtensionInterface`.
 
 ## 4.0.0